requirements.txt
websocket-client
Pillow
requests
yaml